[BE] feat: letter 기능 추가
- stt와 gpt api를 사용한 letter 기능
- S3에 저장은 아직 하는중

+++ BE/project/src/main/java/com/ssafy/project/domain/Letter.java
@@ -22,7 +22,7 @@ public class Letter {
     private Long letterId;
 
     @Column(name = "child_id")
-    private Integer childId;
+    private Long childId;
 
     @Column(name = "letter")
     private String letter;
